[추가] cell에 대해 BaseArchitecture를 도입
